requirements.txt
flask 
flask-cors 
google-generativeai